+++ src/config.h
@@ -24,6 +24,7 @@ typedef struct
     unsigned int skipBackspaceEcho;     // Skip the next incoming character after a backspace from keyboard
     unsigned int swapDelWithBackspace;  // Substitute DEL (0x7F) with BACKSPACE (0x08)
     unsigned int showRC2014Logo;        // Show the RC2014 logo at startup
+    char         keyboardLayout[3];     // Keyboard layout (de,uk,us, ...)
 } tPiGfxConfig;
 
 extern tPiGfxConfig PiGfxConfig;